adding file 102

+++ functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 
 int main(void)
 {
-	long long int i = 0, j = 1, k, sum = 0;
+	long int i = 0, j = 1, k, sum = 0;
 
 	for (k = 1; k <= 50; k++)
 	{
 		sum = i + j;
 		i = j;
 		j = sum;
-		printf("%d, ", sum);
+		printf("%ld, ", sum);
 	}
 
 	return (0);
